sstate: Fix sstate mirrors with BB_FETCH_PREMIRRORONLY
In case BB_FETCH_PREMIRRORONLY is set, the fetcher.download() function
sets BB_NO_NETWORK = "1" after downloading the first file. This alters
the precondition for the second and third loop run, disallowing network
access entirely. As a consequence, the .siginfo and .sig files can
never be downloaded.

Fix this by creating a new copy of localdata for each loop run.

Alternative solutions would be:
- Setting BB_FETCH_PREMIRRORONLY = "0" in localdata
- Setting BB_NO_NETWORK = "0" in each loop run

Both alternatives are highly dependent on the concrete implementation
of the fetcher function, thus making them likely to break in the
future.

Signed-off-by: Julian Haller <julian.haller@philips.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>