rootfs-postcommads: Disable write_image_manifest by default
We now generate SPDX output by default and this is more powerful than any
simple manifest output. The current manifest is also in a hard to parse
format which is something we want to move away from.

Disable the image manifest writing by default going forward for the
above reasons.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>